requirements.txt
flask
flask-sqlalchemy
bcrypt
cryptography
PyJWT
python-dotenv
